+++ lib/mix/tasks/surface/surface.init/project_patchers/layouts.ex
@@ -1,57 +1,50 @@
 defmodule Mix.Tasks.Surface.Init.ProjectPatchers.Layouts do
   @moduledoc false
 
-  alias Mix.Tasks.Surface.Init.FilePatchers
-
   @behaviour Mix.Tasks.Surface.Init.ProjectPatcher
+
+  alias Mix.Tasks.Surface.Init.FilePatchers
 
   @impl true
   def specs(%{layouts: true} = assigns) do
-    %{
-      web_module: web_module,
-      web_path: web_path,
-      web_module_path: web_module_path,
-      tailwind: tailwind?
-    } = assigns
-
-    subfolder = if tailwind?, do: "tailwind", else: "default"
+    %{web_path: web_path, web_module: web_module} = assigns
 
     [
-      {:patch, web_module_path, [add_layout_config_to_view_macro(web_path, web_module)]},
-      {:create, "layouts/#{subfolder}/index.sface", Path.join(web_path, "templates/page")},
-      {:delete, Path.join(web_path, "templates/page/index.html.heex")},
-      {:create, "layouts/#{subfolder}/app.sface", Path.join(web_path, "templates/layout")},
-      {:delete, Path.join(web_path, "templates/layout/app.html.heex")},
-      {:create, "layouts/#{subfolder}/live.sface", Path.join(web_path, "templates/layout")},
-      {:delete, Path.join(web_path, "templates/layout/live.html.heex")},
-      {:create, "layouts/#{subfolder}/root.sface", Path.join(web_path, "templates/layout")},
-      {:delete, Path.join(web_path, "templates/layout/root.html.heex")}
+      {:patch, "#{web_path}/components/layouts.ex", [add_embed_sface_calls_to_layouts(web_module)]},
+      {:create, "layouts/app.sface", Path.join(web_path, "components/layouts")},
+      {:delete, Path.join(web_path, "components/layouts/app.html.heex")},
+      {:create, "layouts/root.sface", Path.join(web_path, "components/layouts")},
+      {:delete, Path.join(web_path, "components/layouts/root.html.heex")}
     ]
   end
 
   def specs(_assigns), do: []
 
-  def add_layout_config_to_view_macro(web_path, web_module) do
+  def add_embed_sface_calls_to_layouts(web_module) do
     %{
-      name: "Configure `Surface.View` in view config",
+      name: "Embed sface templates in the Layouts module",
       patch:
-        &FilePatchers.Phoenix.append_code_to_view_macro(
+        &FilePatchers.Phoenix.append_code_to_layouts_module(
           &1,
-          ~s[use Surface.View, root: "#{web_path}/templates"],
-          "use Surface.View",
-          web_module
+          """
+          embed_sface "layouts/root.sface"
+          embed_sface "layouts/app.sface"\
+          """,
+          "embed_sface",
+          Module.concat(web_module, "Layouts")
         ),
       instructions: """
-      Set up `Surface.View` in your view config so you can uses Surface files as dead views/layouts.
+      If you want to use `.sface` files instead of `.html.heex` files as your layout templates,
+      use the `embed_sface` macro to embed the templates in your Layouts module.
 
       # Example
 
-      ```elixir
-      def view do
-        quote do
-          ...
-          use Surface.View, root: "my_app_web/templates"\
-        end
+      ```
+      defmodule #{inspect(web_module)}.Layouts do
+        use MyAppWeb, :html
+
+        embed_sface "layouts/root.sface"
+        embed_sface "layouts/app.sface"
       end
       ```
       """