+++ test/mix/tasks/surface/surface.init/project_patchers/layouts_test.exs
@@ -4,65 +4,40 @@ defmodule Mix.Tasks.Surface.Init.ProjectPatchers.LayoutsTest do
   alias Mix.Tasks.Surface.Init.Patcher
   import Mix.Tasks.Surface.Init.ProjectPatchers.Layouts
 
-  describe "add_layout_config_to_view_macro" do
+  describe "add_embed_sface_calls_to_layouts" do
     test "add layout config to view macro" do
       code = """
-      defmodule MyAppWeb do
-        def view do
-          quote do
-            use Phoenix.View,
-              root: "lib/my_app_web/templates",
-              namespace: MyAppWeb
+      defmodule MyAppWeb.Layouts do
+        use MyAppWeb, :html
 
-            # Include shared imports and aliases for views
-            unquote(view_helpers())
-            import Surface
-          end
-        end
+        embed_templates "layouts/*"
       end
       """
 
-      {:patched, updated_code} =
-        Patcher.patch_code(code, add_layout_config_to_view_macro("lib/my_app_web", MyAppWeb))
+      {:patched, updated_code} = Patcher.patch_code(code, add_embed_sface_calls_to_layouts(MyAppWeb))
 
       assert updated_code == """
-             defmodule MyAppWeb do
-               def view do
-                 quote do
-                   use Phoenix.View,
-                     root: "lib/my_app_web/templates",
-                     namespace: MyAppWeb
+             defmodule MyAppWeb.Layouts do
+               use MyAppWeb, :html
 
-                   # Include shared imports and aliases for views
-                   unquote(view_helpers())
-                   import Surface
-                   use Surface.View, root: "lib/my_app_web/templates"
-                 end
-               end
+               embed_templates "layouts/*"
+               embed_sface "layouts/root.sface"
+               embed_sface "layouts/app.sface"
              end
              """
     end
 
     test "don't apply it if already patched" do
       code = """
-      defmodule MyAppWeb do
-        def view do
-          quote do
-            use Phoenix.View,
-              root: "lib/my_app_web/templates",
-              namespace: MyAppWeb
+      defmodule MyAppWeb.Layouts do
+        use MyAppWeb, :html
 
-            # Include shared imports and aliases for views
-            unquote(view_helpers())
-            import Surface
-            use Surface.View, root: "lib/my_app_web/templates"
-          end
-        end
+        embed_templates "layouts/*"
+        embed_sface "layouts/root.sface"
       end
       """
 
-      assert {:already_patched, ^code} =
-               Patcher.patch_code(code, add_layout_config_to_view_macro("lib/my_app_web", MyAppWeb))
+      assert {:already_patched, ^code} = Patcher.patch_code(code, add_embed_sface_calls_to_layouts(MyAppWeb))
     end
   end
 end