+++ test/support/mix/tasks/compile/surface_test/fake_link.ex
@@ -4,6 +4,8 @@ defmodule Mix.Tasks.Compile.SurfaceTest.FakeLink do
   def render(assigns) do
     ~F"""
     <style>
+      @import "./fake_link.css";
+
       .link { padding: 10px; }
     </style>
     FAKE LINK